infrastructure/monitoring/metrics-collector/requirements.txt
azure-functions
azure-identity
azure-monitor-query
azure-cosmos
